Quick Fix

+++ lib/models/ingredient.dart
@@ -4,10 +4,10 @@ import 'package:noorish_app/hooks/barcode_api.dart';
 @JsonSerializable()
 class Ingredient {
   final String name = '';
-  double protein = 0;
-  double carbs = 0;
-  double calories = 0;
-  double fats = 0;
+  int protein = 0;
+  int carbs = 0;
+  int calories = 0;
+  int fats = 0;
   final String barcode;
 
   Ingredient({required this.barcode});
